统一db 枚举表相关
# message auto-generated for no-merge-commit merge:
merge linux into master

统一db 枚举表相关

Created-by: msprofDeveloper

Match-id-7f7a1a6bca04a12003d093aa8d9a64aad05264bb

+++ analysis/csrc/parser/environment/context.h
@@ -52,7 +52,7 @@ public:
     // 获取start_info end_info中的时间
     bool GetProfTimeRecordInfo(Utils::ProfTimeRecord &record, const std::string &profPath = "");
     // 返回info.json 中的pid
-    uint64_t GetPidFromInfoJson(uint16_t deviceId = DEFAULT_DEVICE_ID, const std::string &profPath = "");
+    uint32_t GetPidFromInfoJson(uint16_t deviceId = DEFAULT_DEVICE_ID, const std::string &profPath = "");
     // 返回samplejson.json 中的msprofBinPid
     int64_t GetMsBinPid(const std::string &profPath);
     // 获取start_log中的相关时间
